+++ Razor.Blade/Markup/Tag/TagText.cs
@@ -1,10 +1,12 @@
 ﻿namespace ToSic.Razor.Markup
 {
-    public class TagText: TagBase
+    public class TagText: Tag<TagText>
     {
-        public TagText(string text) : base(tagOverride: text)
-        {
+        public TagText(string text) : base(true, tagOverride: text) { }
 
-        }
+        private TagText(TagText original, CloneChanges changes) : base(original, changes) { }
+
+        /// <inheritdoc />
+        internal override TagText CwC(CloneChanges changes) => new TagText(this, changes);
     }
 }